requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
pyyaml>=6.0
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
youtube-transcript-api>=0.6.2
pytest>=8.0.0
ruff>=0.9.0
